==> snpnet/1_phe_prep.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(data.table))


### input files

gwas_covar_f <- '@@@@@@@@@@@@@@@@@@@'


biomarker_f <- '@@@@@@@@@@@@@@@@@@@'


biomarker_trait_list_f <- '@@@@@@@@@@@@@@@@@@@'


master_phe_f <- '@@@@@@@@@@@@@@@@@@@'


### output

phe_info_f <- 'pheno.info.tsv'
phe_file <- 'data/pheno.tsv'


gwas_covar_df <- fread(
    gwas_covar_f,
    select=c('FID', 'IID', 'population', 'split', 'age', 'sex', 'N_CNV', 'LEN_CNV', 'Array', paste0('PC', 1:10)),
    colClasses=c('FID'='character', 'IID'='character')
)


biomarker_trait_list_df <- fread(biomarker_trait_list_f)


biomarker_df <- fread(
    biomarker_f,
    select=c(
        'FID', 'IID', 
        biomarker_trait_list_df %>% 
        filter(annotation != 'Fasting_glucose') %>%
        select(annotation) %>% pull()
    ),
    colClasses=c('FID'='character', 'IID'='character')
)


biomarker_df %>% dim()

blood_ukb_fields <- c(
    # http://biobank.ctsu.ox.ac.uk/crystal/label.cgi?id=100081
30160,
30220,
30150,
30210,
30030,
30020,
30300,
30290,
30280,
30120,
30180,
30050,
30060,
30040,
30100,
30260,
30270,
30130,
30190,
30140,
30200,
30170,
30230,
30080,
30090,
30110,
30010,
30070,
30250,
30240,
30000
)


master_phe_df <- fread(
    master_phe_f,
    select=c(
        'FID', 'IID', 
        paste0('INI', sort(blood_ukb_fields))
    ),
    colClasses=c('FID'='character', 'IID'='character')
)



master_phe_df %>% dim()

library(googlesheets)
gs_auth(token = "@@@@@@@@@@@@@@@@@@@")

file <- '@@@@@@@@@@@@@@@@@@@'
GBE_names_df <- file %>% gs_url() %>% gs_read(ws = 'GBE_names')


phe_meta_df <- 
GBE_names_df %>% 
filter(GBE_ID %in% paste0('INI', sort(blood_ukb_fields))) %>%
select(GBE_ID, GBE_NAME, GBE_short_name) %>%
rename('pheno_plot' = 'GBE_short_name', 'pheno' = 'GBE_NAME') %>%
bind_rows(
    biomarker_trait_list_df %>% 
    filter(annotation != 'Fasting_glucose') %>%
    rename('pheno' = 'annotation', 'pheno_plot' = 'name') %>%
    mutate(GBE_ID = pheno)
) %>%
select(GBE_ID, pheno, pheno_plot)


phe_meta_df %>% colnames()

phe_meta_df %>% 
rename('#GBE_ID' = 'GBE_ID') %>%
fwrite(phe_info_f, sep='\t')


merged_df <- gwas_covar_df %>%
filter(population == 'white_british') %>%
left_join(
    biomarker_df, by=c('FID', 'IID')
) %>%
left_join(
    master_phe_df, by=c('FID', 'IID')
)


merged_df %>% colnames()

merged_df %>% fwrite(phe_file, sep='\t')
